+++ assets/scss/components/foundation/modal/_modal.scss
@@ -131,3 +131,7 @@
         }
     }
 }
+
+.hide-content {
+    opacity: 0;
+}
