+++ templates/components/cart/content.html
@@ -53,7 +53,7 @@
                             {{/each}}
                         </dl>
 
-                        <a href="#" data-item-edit="{{id}}">{{lang 'cart.checkout.change'}}</a>
+                        <a href="#" data-item-edit="{{id}}" data-product-id="{{product_id}}">{{lang 'cart.checkout.change'}}</a>
                     {{/if}}
 
                     {{#if type '==' 'GiftCertificate'}}
