+++ templates/components/cart/modals/configure-product.html
@@ -1,5 +1,11 @@
 <div class="modal-header">
-    <h2 class="modal-header-title">{{lang 'cart.reconfigure_product' name=product_name}}</h2>
+    <h2 
+        class="modal-header-title" 
+        data-product-variant 
+        data-name="{{product_name}}"
+        >
+        {{lang 'cart.reconfigure_product' name=product_name}}
+    </h2>
 </div>
 <div class="modal-body" data-product-option-change>
     <div class="alertMessageBox" hidden>
@@ -9,11 +15,11 @@
     <form id="CartEditProductFieldsForm" method="post" action="cart.php" enctype="multipart/form-data">
         <input type="hidden" name="action" value="EditProductFieldsInCart">
         <input type="hidden" name="item_id" value="{{quote_item_id}}">
-
+        <div data-product-attributes-wrapper>
         {{#each options}}
             {{{dynamicComponent 'components/products/options'}}}
         {{/each}}
-
+        </div>
         <div class="form-actions">
             <input class="button button--primary" type="submit" value="{{lang 'common.save'}}">
         </div>
